# docs/api.yaml
openapi: 3.0.3
info:
  title: Campus Team Platform API
  version: 1.0.0
  description: 校园活动组队平台 API 文档
servers:
  - url: http://localhost:8080
paths:
...
  /api/join/request:
    post:
      summary: 发送加入请求
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRequestCreate'
      responses:
        '200':
          description: 请求已发送
...
  /api/team/{id}/join-requests:
    get:
      summary: 获取团队加入请求
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 请求列表
  /api/join/{requestId}/approve:
    post:
      summary: 批准加入请求
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 已批准
  /api/join/{requestId}/reject:
    post:
      summary: 拒绝加入请求
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 已拒绝
...
components:
  schemas:
...
    JoinRequestCreate:
      type: object
      properties:
        teamId:
          type: integer
        reason:
          type: string
...